[2-lab] update `self_study_hours` data type (`int` -> `long`); fix problem where `self_study_hours` wasn't available to sort by

+++ common/src/main/java/ru/xgodness/endpoint/faculties/dto/Discipline.java
@@ -11,5 +11,5 @@ import lombok.*;
 public class Discipline {
     private String faculty;
     private String name;
-    private int selfStudyHours;
+    private long selfStudyHours;
 }
